ml_inference_server/requirements.txt

fastapi==0.109.0
uvicorn==0.27.0
torch>=2.2.0
transformers>=4.41.0
python-multipart==0.0.6
numpy<2.0
pillow==10.2.0
soundfile==0.12.1
librosa==0.10.1

